app/api/spotify/serializers.py
from rest_framework import serializers
from backend.models import Query

class QuerySerializer(serializers.ModelSerializer):
    class Meta:
        model = Query
        fields = ['id', 'name', 'parameters', 'created_at', 'updated_at']
        read_only_fields = ['id', 'created_at', 'updated_at']

    def create(self, validated_data):
        user = self.context['request'].user
        return Query.objects.create(user=user, **validated_data)
