app/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
requests==2.31.0

